+ Split boundary conditions into macroscopic and microscopic parts + Implemented a stress free BC but its currently copying post-streamed PDF's when it should be copying post-collision PDF's + Steady State Check adjusted so that residual is computed from total energy, so it is no longer restricted to x-aligned flows + Added y-plane and z-plane zhou/he pressure conditions

+++ LBM-C/model_builder.cu
@@ -220,7 +220,7 @@ class ModelBuilder
 			}
 
 			input_handler.read_field(domain_h->rho, "Rho");
-			cudasafe(cudaMemcpy(rho_d, rho_h, sizeof(int)*domain_size,cudaMemcpyHostToDevice),"Model Builder: Copy to device memory failed!");
+			cudasafe(cudaMemcpy(rho_d, rho_h, sizeof(double)*domain_size,cudaMemcpyHostToDevice),"Model Builder: Copy to device memory failed!");
 		}
 
 		if(domain_constants_h->init_type == 0)
